fix(native): resolve Rust compile error and address review feedback
- Fix type mismatch in build_pipeline.rs: change FAST_PATH_MIN_EXISTING_FILES
  from usize to i64 to match get_existing_file_count return type
- Remove misleading underscore prefix from _rel_path variable that is
  actually used (build_pipeline.rs:507)
- Replace missed hardcoded 120 literal with DATAFLOW_TRUNCATION_LIMIT
  constant at dataflow.rs:1071

+++ crates/codegraph-core/src/constants.rs
@@ -27,4 +27,5 @@ pub const DATAFLOW_TRUNCATION_LIMIT: usize = 120;
 pub const FAST_PATH_MAX_CHANGED_FILES: usize = 5;
 
 /// Minimum existing file count required before the fast path is considered.
-pub const FAST_PATH_MIN_EXISTING_FILES: usize = 20;
+/// Typed as `i64` to match the SQLite row-count return type from `get_existing_file_count`.
+pub const FAST_PATH_MIN_EXISTING_FILES: i64 = 20;
